Moved perform_near_match() function to near_matching module
