+++ collatex-pythonport/collatex/display_module.py
@@ -64,15 +64,18 @@ def display_variant_graph_as_SVG(graph,svg_output):
             reading = ("<TR><TD ALIGN='LEFT'><FONT FACE='Bukyvede'>{}</FONT></TD><TD ALIGN='LEFT'>{}</TD></TR>").format(key,', '.join(value))
             readings.append(reading)
         a.add_node(mapping[n], label='<<TABLE CELLSPACING="0">' + "".join(readings) + '</TABLE>>')
-    # add edges
+    # add regular (token sequence) edges
     for u,v,edgedata in graph.graph.edges_iter(data=True):
         # print('regular edges ', u, v, edgedata)
         label = edgedata['label']
         a.add_edge(str(mapping[u]), str(mapping[v]), label=label)
+    # add near-match edges
+    # TODO: Show all near edges (currently), or just the top one?
     for u,v,edgedata in graph.near_graph.edges_iter(data=True):
         # print('near-match edges ', u, v, edgedata)
         label = str('{:3.2f}'.format(edgedata['weight']))
         a.add_edge(str(mapping[u]), str(mapping[v]), style='dashed', label=label)
+    # Add rank='same' information
     for key, value in ranking.byRank.items():
         a.add_subgraph([mapping[item] for item in value], rank='same')
     # display using the IPython SVG module
